requirements.txt
gradio==5.0.1
biopython